add getLightSource fetch logic in repostory

+++ light_master_app/lib/modules/dashboard/models/light.dart
@@ -14,8 +14,8 @@ class SolidLight extends Light {
 
 class EffectLight extends Light {
   final String effect;
-  final double brightness;
-  final double speed;
+  final int brightness;
+  final int speed;
 
   EffectLight(this.effect, this.brightness, this.speed);
 }
